server/requirements.txt
aniso8601==10.0.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dateparser==1.2.1
dnspython==2.6.1
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
flask-restx==1.1.0
Flask-Session==0.5.0
fonttools==4.56.0
fsspec==2025.3.0
gdown==5.1.0
google-api-core==2.24.2
google-api-python-client==2.123.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.69.1
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==24.2
pickle-mixin==1.0.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
proto-plus==1.26.1
protobuf==5.29.3
py7zr==0.20.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
rarfile==4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.5.1
six==1.17.0
spacy==3.7.4
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==2.3.7
zipp==3.20.2
gunicorn
waitress
openai
nltk
seaborn
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz